refactor(leads): the promotion's event payloads move to promoteevents.go
promote.go crossed the 500-line ceiling by four lines; the two wire-shape
builders are their own concept.
